+++ domain/domain-connector/src/main/kotlin/by/miaskor/domain/connector/ClientConnector.kt
@@ -20,5 +20,5 @@ interface ClientConnector {
   fun createClient(clientDtoRequest: ClientDtoRequest): ClientDtoResponse
 
   @RequestLine("POST /byBotId/{bot_id}")
-  fun getClientByBotId(@Param("bot_id") botId: Int): ClientDtoResponse
+  fun getClientByBotId(@Param("bot_id") botId: Long): ClientDtoResponse
 }
